Rename and clean localdecl handling
LocalDeclarations -> LocalNodeDeclarations (as it returns ast nodes)
LocalDecls -> LocalNameDeclarations (as it returns names)

And provide both from the same file!
